pyproject.toml
[tool.poetry]
name = "antgirl"
version = "0.1.0"
description = ""
authors = ["Harry <harry@apartalk.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.77.1"
alembic = "^1.7.7"
aiomysql = "^0.1.1"
dependency-injector = "^4.39.1"
PyJWT = "^2.3.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
PyMySQL = "^1.0.2"
Scrapy = "^2.6.1"
celery = "^5.2.6"
flower = "^1.0.0"
redis = "^4.3.1"
scrapy-fake-useragent = "^1.4.4"
xmltodict = "^0.13.0"
PyPubSub = "^4.0.3"

[tool.poetry.dev-dependencies]
aiosqlite = "^0.17.0"
httpx = "^0.22.0"
pytest-factoryboy = "^2.2.1"
pytest = "^7.1.2"
pytest-xdist = "^2.5.0"
pytest-asyncio = "^0.18.3"
pyright = "^1.1.245"
black = "^22.3.0"
asgi-lifespan = "^1.0.1"
Faker = "^13.11.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
